+++ src/components/super-admin/company/index.ts
@@ -1,0 +1,3 @@
+export { CompanyMiniCard } from './CompanyMiniCard'
+export { CompanyMacroCard } from './CompanyMacroCard'
+export { CompanyStatusBadge } from './CompanyStatusBadge'